Generalized EC and BN managers to accept client interfaces, created IBeaconApiProvider instead of forcing HTTP
