+++ node/services/bn-manager.go
@@ -8,7 +8,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/rocket-pool/node-manager-core/api/types"
 	"github.com/rocket-pool/node-manager-core/beacon"
-	"github.com/rocket-pool/node-manager-core/beacon/client"
 )
 
 // This is a proxy for multiple Beacon clients, providing natural fallback support if one of them fails.
@@ -21,14 +20,7 @@ type BeaconClientManager struct {
 }
 
 // Creates a new BeaconClientManager instance
-func NewBeaconClientManager(primaryProvider string, fallbackProvider string, chainID uint, clientTimeout time.Duration) (*BeaconClientManager, error) {
-	var primaryBc beacon.IBeaconClient
-	var fallbackBc beacon.IBeaconClient
-	primaryBc = client.NewStandardHttpClient(primaryProvider, clientTimeout)
-	if fallbackProvider != "" {
-		fallbackBc = client.NewStandardHttpClient(fallbackProvider, clientTimeout)
-	}
-
+func NewBeaconClientManager(primaryBc beacon.IBeaconClient, fallbackBc beacon.IBeaconClient, chainID uint, clientTimeout time.Duration) (*BeaconClientManager, error) {
 	return &BeaconClientManager{
 		primaryBc:       primaryBc,
 		fallbackBc:      fallbackBc,
