+++ node/services/ec-manager.go
@@ -10,17 +10,14 @@ import (
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/ethclient"
 	apitypes "github.com/rocket-pool/node-manager-core/api/types"
 	"github.com/rocket-pool/node-manager-core/eth"
 )
 
 // This is a proxy for multiple ETH clients, providing natural fallback support if one of them fails.
 type ExecutionClientManager struct {
-	primaryEcUrl    string
-	fallbackEcUrl   string
-	primaryEc       *ethclient.Client
-	fallbackEc      *ethclient.Client
+	primaryEc       eth.IExecutionClient
+	fallbackEc      eth.IExecutionClient
 	primaryReady    bool
 	fallbackReady   bool
 	expectedChainID uint
@@ -28,24 +25,8 @@ type ExecutionClientManager struct {
 }
 
 // Creates a new ExecutionClientManager instance
-func NewExecutionClientManager(primaryEcUrl string, fallbackEcUrl string, chainID uint, clientTimeout time.Duration) (*ExecutionClientManager, error) {
-	primaryEc, err := ethclient.Dial(primaryEcUrl)
-	if err != nil {
-		return nil, fmt.Errorf("error connecting to primary EC at [%s]: %w", primaryEcUrl, err)
-	}
-
-	// Get the fallback EC url, if applicable
-	var fallbackEc *ethclient.Client
-	if fallbackEcUrl != "" {
-		fallbackEc, err = ethclient.Dial(fallbackEcUrl)
-		if err != nil {
-			return nil, fmt.Errorf("error connecting to fallback EC at [%s]: %w", fallbackEcUrl, err)
-		}
-	}
-
+func NewExecutionClientManager(primaryEc eth.IExecutionClient, fallbackEc eth.IExecutionClient, chainID uint, clientTimeout time.Duration) (*ExecutionClientManager, error) {
 	return &ExecutionClientManager{
-		primaryEcUrl:    primaryEcUrl,
-		fallbackEcUrl:   fallbackEcUrl,
 		primaryEc:       primaryEc,
 		fallbackEc:      fallbackEc,
 		primaryReady:    true,
@@ -254,6 +235,16 @@ func (m *ExecutionClientManager) SyncProgress(ctx context.Context) (*ethereum.Sy
 	})
 }
 
+/// =======================
+/// ChainIDReader Functions
+/// =======================
+
+func (m *ExecutionClientManager) ChainID(ctx context.Context) (*big.Int, error) {
+	return runFunction1(m, ctx, func(client eth.IExecutionClient) (*big.Int, error) {
+		return client.ChainID(ctx)
+	})
+}
+
 /// =================
 /// Manager Functions
 /// =================
@@ -293,7 +284,7 @@ func (m *ExecutionClientManager) CheckStatus(ctx context.Context, checkChainIDs 
 }
 
 // Check the client status
-func checkEcStatus(ctx context.Context, client *ethclient.Client, checkChainIDs bool) apitypes.ClientStatus {
+func checkEcStatus(ctx context.Context, client eth.IExecutionClient, checkChainIDs bool) apitypes.ClientStatus {
 	status := apitypes.ClientStatus{}
 
 	if checkChainIDs {
